Fixed IllegalStateException
The plugin previously only stored one list with one of each title as a BukkitRunnable, and scheduling the same instance twice causes problems. Fixed this by adding a clone function in Title that simply returns a new instance of the runnable with the same settings, and then re-creates the entire list each time the display is needed. This isn't the best solution because it relies on the garbage collector but it is the simplest and fastest fix to this problem without rewriting the code. Also, my eclipse has a formatter, and so most of the little changes are from that.